+++ test/support/bin/test_bins.ex
@@ -125,7 +125,7 @@ defmodule Membrane.Support.Bin.TestBins do
     end
 
     @impl true
-    def handle_pad_added(Pad.ref(:input, _) = pad, _ctx, state) do
+    def handle_pad_added(Pad.ref(:input, _id) = pad, _ctx, state) do
       links = [
         link_bin_input(pad) |> to(:filter)
       ]
@@ -133,7 +133,7 @@ defmodule Membrane.Support.Bin.TestBins do
       {{:ok, notify: {:handle_pad_added, pad}, spec: %ParentSpec{links: links}}, state}
     end
 
-    def handle_pad_added(Pad.ref(:output, _) = pad, _ctx, state) do
+    def handle_pad_added(Pad.ref(:output, _id) = pad, _ctx, state) do
       links = [
         link(:filter) |> to_bin_output(pad)
       ]
@@ -175,7 +175,7 @@ defmodule Membrane.Support.Bin.TestBins do
     end
 
     @impl true
-    def handle_pad_added(Pad.ref(:input, _) = pad, _ctx, state) do
+    def handle_pad_added(Pad.ref(:input, _id) = pad, _ctx, state) do
       links = [
         link_bin_input(pad) |> to(:filter1)
       ]
@@ -183,7 +183,7 @@ defmodule Membrane.Support.Bin.TestBins do
       {{:ok, notify: {:handle_pad_added, pad}, spec: %ParentSpec{links: links}}, state}
     end
 
-    def handle_pad_added(Pad.ref(:output, _) = pad, _ctx, state) do
+    def handle_pad_added(Pad.ref(:output, _id) = pad, _ctx, state) do
       links = [
         link(:filter2) |> to_bin_output(pad)
       ]